+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/clientSide.tsx
@@ -40,10 +40,6 @@ export const FormWrapper = ({
   } = useGCFormsContext();
   const [captchaFail, setCaptchaFail] = useState(false);
   const { cachedSession } = useResponsesCache();
-  const captchaToken = React.useRef("");
-  // TODO : If the formRecord contains file inputs Save and Resume is not available
-  const resetCaptchaRef = React.useRef<{ resetToken: () => void }>({ resetToken: () => {} });
-
   const saveAndResume = formRecord?.saveAndResume;
 
   // Generate form elements on the client to ensure Formik context is available
@@ -137,8 +133,6 @@ export const FormWrapper = ({
         currentGroup={currentGroup}
         setCaptchaFail={setCaptchaFail}
         captchaFail={captchaFail}
-        captchaToken={captchaToken}
-        resetCaptchaRef={resetCaptchaRef}
       >
         {currentForm}
       </Form>