+++ components/clientComponents/forms/Form/SubmitButton.tsx
@@ -4,9 +4,10 @@ import { useTranslation } from "@i18n/client";
 
 interface SubmitButtonProps {
   disabled: boolean;
+  isSubmitting?: boolean;
   submissionError?: boolean;
 }
-export const SubmitButton = ({ disabled, submissionError }: SubmitButtonProps) => {
+export const SubmitButton = ({ disabled, isSubmitting, submissionError }: SubmitButtonProps) => {
   const { t } = useTranslation();
   // Used to show a spinner for the initial validation and hCAPTCHA loading steps
   const [loading, setLoading] = useState(false);
@@ -15,7 +16,7 @@ export const SubmitButton = ({ disabled, submissionError }: SubmitButtonProps) =
     <Button
       id="form-submit-button"
       type="submit"
-      disabled={disabled}
+      disabled={disabled || isSubmitting}
       onClick={() => {
         setLoading(true);
       }}