+++ packages/hcaptcha/src/client/index.ts
@@ -1,3 +1,5 @@
+export { HCaptchaForm } from "./HCaptchaForm";
+export type { HCaptchaFormHandle, HCaptchaFormProps } from "./HCaptchaForm";
 export { useHCaptcha } from "./useHCaptcha";
 export type {
   HCaptchaExecutionResult,