+++ packages/hcaptcha/CHANGELOG.md
@@ -5,15 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-09-03
+
+- Adds a reusable React form wrapper
+
+## [1.0.3] - 2026-09-02
+
+- Adds more error cases to the retry loop and updates the readme
+
 ## [1.0.2] - 2026-09-01
 
 - Testing a change for the workflow
 
 ## [1.0.1] - 2026-09-01
 
-- simplify the README
-
+- Simplify the README
 
 ## [1.0.0] - 2026-09-01
 
-- moved hCAPTCHA related client and server code to this internal package
+- Moved hCaptcha client and server code into the internal `@gcforms/hcaptcha` package